+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to the "quick-ctags-parser" extension will be documented in 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [v1.0.4]
+
+- Fix multi workspace
+- Fix absolute file paths
 
 ## [v1.0.1]